Tie Composer dependency versions to the versions developed against
When installing a package as a library, Composer doesn't respect the lock
file in the package. This meant that this library could be installed with
a version of Guzzle that it wouldn't work with, if the requiring project
already required an incompatible version of Guzzle.

This lock may be a little too restrictive, but I'm not familiar enough with
Guzzle to know at a glance which API versions this project is compatible with.